Conflicts:
	src/UI/BanHangForm.form
	src/UI/BanHangForm.java
	src/UI/DangNhapFrom.form
	src/UI/DangNhapFrom.java
	src/UI/GiamGiaForm.form
	src/UI/GiamGiaForm.java
	src/UI/HTNSJFrame.form
	src/UI/HTNSJFrame.java
	src/UI/NhanVienForm.form
	src/UI/NhanVienForm.java
	src/UI/SanPhamForm.form
	src/UI/SanPhamForm.java
	src/UI/TaiKhoanForm.form
	src/UI/TaiKhoanForm.java
	src/UI/ThongKeForm.form
	src/UI/ThongKeForm.java

+++ OneDrive/Tài liệu/NetBeansProjects/CafeHTNS/src/Model/ThongKe.java
@@ -12,12 +12,12 @@ public class ThongKe {
     private String maTk;
     private String maHD;
     private String ngayTk;
-    private float luong;
+    private double luong;
 
     public ThongKe() {
     }
 
-    public ThongKe(String maTk, String maHD, String ngayTk, float luong) {
+    public ThongKe(String maTk, String maHD, String ngayTk, double luong) {
         this.maTk = maTk;
         this.maHD = maHD;
         this.ngayTk = ngayTk;
@@ -48,11 +48,11 @@ public class ThongKe {
         this.ngayTk = ngayTk;
     }
 
-    public float getLuong() {
+    public double getLuong() {
         return luong;
     }
 
-    public void setLuong(float luong) {
+    public void setLuong(double luong) {
         this.luong = luong;
     }
